requirements.txt
streamlit~=1.40.1
requests~=2.32.3
pyzbar~=0.1.9
opencv-python-headless
streamlit_qrcode_scanner
opencv-python~=4.10.0.84
python-dotenv~=1.0.1
pandas~=2.2.3